#23 extract dbReadEvents method
